requirements.txt
Django==1.8.4
Pillow==2.9.0
PyPDF2==1.25.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.2.3
gunicorn==19.3.0
html5lib==0.999999
psycopg2==2.6.1
reportlab==3.2.0
requests==2.7.0
simplejson==3.8.0
six==1.9.0
static3==0.6.1
whitenoise==2.0.3
wsgiref==0.1.2
xhtml2pdf==0.0.6
